feat: update scores across portal
